+++ whitelist.py
@@ -1,4 +1,4 @@
 _.fore  # unused attribute (tests/_test.py:51)
 fixture_color  # unused function (tests/conftest.py:9)
 fixture_populated_colors  # unused function (tests/conftest.py:15)
-_.populate_colors  # unused method (object_colors/__init__.py:139)
+_.populate_colors  # unused method (object_colors/__init__.py:158)